# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * බෝdima.lk Boarding Finder API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type RegisterInputRole = typeof RegisterInputRole[keyof typeof RegisterInputRole];


export const RegisterInputRole = {
  student: 'student',
  owner: 'owner',
} as const;

export interface RegisterInput {
  email: string;
  /** @minLength 6 */
  password: string;
  first_name: string;
  last_name: string;
  /** @nullable */
  phone?: string | null;
  role: RegisterInputRole;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface UserProfile {
  id: number;
  email: string;
  first_name: string;
  last_name: string;
  /** @nullable */
  phone?: string | null;
  role: string;
  is_active?: boolean;
  created_at: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface UserProfileUpdate {
  first_name?: string;
  last_name?: string;
  phone?: string;
}

export interface Listing {
  id: number;
  name: string;
  area: string;
  location?: string;
  price: number;
  type?: string;
  gender?: string;
  rating: number;
  review_count: number;
  img: string;
  amenities: string[];
  /** @nullable */
  badge?: string | null;
  status: string;
  /** @nullable */
  desc?: string | null;
}

export interface ListingImage {
  id: number;
  url: string;
  /** @nullable */
  caption?: string | null;
  is_primary?: boolean;
}

export interface ListingDetail {
  id: number;
  name: string;
  area: string;
  location?: string;
  price: number;
  type?: string;
  gender?: string;
  rating: number;
  review_count: number;
  img: string;
  amenities: string[];
  /** @nullable */
  badge?: string | null;
  status: string;
  /** @nullable */
  desc?: string | null;
  deposit_months?: number;
  meals_included?: boolean;
  wifi_included?: boolean;
  available_from?: string;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  owner_id?: number;
  /** @nullable */
  owner_name?: string | null;
  /** @nullable */
  owner_since?: string | null;
  images?: ListingImage[];
  created_at?: string;
}

export interface ListingsSearchResult {
  listings: Listing[];
  total: number;
}

export interface ListingInput {
  name: string;
  area: string;
  location: string;
  price: number;
  type?: string;
  gender?: string;
  desc: string;
  amenities?: string[];
  deposit_months?: number;
  meals_included?: boolean;
  wifi_included?: boolean;
  available_from?: string;
  img?: string;
  /** @nullable */
  badge?: string | null;
}

export interface ListingUpdate {
  name?: string;
  area?: string;
  location?: string;
  price?: number;
  type?: string;
  gender?: string;
  desc?: string;
  amenities?: string[];
  deposit_months?: number;
  meals_included?: boolean;
  wifi_included?: boolean;
  available_from?: string;
  img?: string;
  /** @nullable */
  badge?: string | null;
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  rejected: 'rejected',
  cancelled: 'cancelled',
} as const;

export interface Booking {
  id: number;
  listing_id: number;
  /** @nullable */
  listing_name?: string | null;
  /** @nullable */
  listing_area?: string | null;
  /** @nullable */
  listing_price?: number | null;
  /** @nullable */
  listing_img?: string | null;
  student_id: number;
  /** @nullable */
  student_name?: string | null;
  /** @nullable */
  student_email?: string | null;
  /** @nullable */
  owner_id?: number | null;
  /** @nullable */
  move_in_date?: string | null;
  /** @nullable */
  message?: string | null;
  status: BookingStatus;
  created_at: string;
}

export interface BookingInput {
  listing_id: number;
  /** @nullable */
  move_in_date?: string | null;
  /** @nullable */
  message?: string | null;
}

export type BookingStatusUpdateStatus = typeof BookingStatusUpdateStatus[keyof typeof BookingStatusUpdateStatus];


export const BookingStatusUpdateStatus = {
  confirmed: 'confirmed',
  rejected: 'rejected',
  cancelled: 'cancelled',
} as const;

export interface BookingStatusUpdate {
  status: BookingStatusUpdateStatus;
}

export interface Review {
  id: number;
  listing_id: number;
  /** @nullable */
  listing_name?: string | null;
  student_id?: number;
  /** @nullable */
  student_name?: string | null;
  /** @nullable */
  student_avatar?: string | null;
  rating: number;
  /** @nullable */
  comment?: string | null;
  created_at: string;
}

export interface ReviewInput {
  listing_id: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  /** @nullable */
  comment?: string | null;
}

export interface Conversation {
  id: number;
  listing_id: number;
  /** @nullable */
  listing_name?: string | null;
  /** @nullable */
  listing_img?: string | null;
  student_id: number;
  /** @nullable */
  student_name?: string | null;
  owner_id: number;
  /** @nullable */
  owner_name?: string | null;
  /** @nullable */
  last_message?: string | null;
  /** @nullable */
  last_message_at?: string | null;
  unread_count?: number;
  created_at: string;
}

export interface Message {
  id: number;
  conversation_id: number;
  sender_id: number;
  /** @nullable */
  sender_name?: string | null;
  body: string;
  is_read?: boolean;
  created_at: string;
}

export interface ConversationDetail {
  id: number;
  listing_id: number;
  /** @nullable */
  listing_name?: string | null;
  /** @nullable */
  listing_img?: string | null;
  student_id: number;
  /** @nullable */
  student_name?: string | null;
  owner_id: number;
  /** @nullable */
  owner_name?: string | null;
  messages: Message[];
  created_at: string;
}

export interface MessageInput {
  body: string;
}

export interface ConversationInput {
  listing_id: number;
}

export interface UnreadCount {
  count: number;
}

export interface AdminUser {
  id: number;
  email: string;
  first_name: string;
  last_name: string;
  /** @nullable */
  phone?: string | null;
  role: string;
  is_active: boolean;
  listing_count?: number;
  created_at: string;
}

export type AdminListingStatusUpdateStatus = typeof AdminListingStatusUpdateStatus[keyof typeof AdminListingStatusUpdateStatus];


export const AdminListingStatusUpdateStatus = {
  approved: 'approved',
  rejected: 'rejected',
  inactive: 'inactive',
} as const;

export interface AdminListingStatusUpdate {
  status: AdminListingStatusUpdateStatus;
  /** @nullable */
  notes?: string | null;
}

export interface AdminUserStatusUpdate {
  is_active: boolean;
}

export interface DashboardStats {
  total_listings: number;
  total_students: number;
  satisfaction_rate: number;
  total_reviews?: number;
}

export interface OwnerSummary {
  total_listings: number;
  active_listings: number;
  pending_bookings: number;
  total_bookings: number;
  confirmed_bookings?: number;
  unread_messages: number;
  average_rating?: number;
  total_reviews?: number;
}

export interface AdminSummary {
  total_users: number;
  total_students?: number;
  total_owners?: number;
  total_listings: number;
  approved_listings?: number;
  pending_listings: number;
  total_bookings: number;
  total_reviews: number;
}

export type SearchListingsParams = {
q?: string;
location?: string;
type?: string;
gender?: string;
min_price?: number;
max_price?: number;
amenities?: string;
sort?: string;
min_rating?: number;
};

export type GetReviewsParams = {
listing_id: number;
};

export type GetAllReviewsParams = {
limit?: number;
};

export type AdminGetListingsParams = {
status?: string;
};

export type AdminGetUsersParams = {
role?: string;
q?: string;
};