# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: බෝdima.lk Boarding Finder API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /listings/search:
    get:
      operationId: searchListings
      tags: [listings]
      summary: Search listings with filters
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: gender
          in: query
          schema:
            type: string
        - name: min_price
          in: query
          schema:
            type: integer
        - name: max_price
          in: query
          schema:
            type: integer
        - name: amenities
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
        - name: min_rating
          in: query
          schema:
            type: number
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingsSearchResult"
  /listings/{id}:
    get:
      operationId: getListing
      tags: [listings]
      summary: Get listing detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── OWNER LISTINGS ────────────────────────────────────────────────
  /owner/listings:
    get:
      operationId: getOwnerListings
      tags: [listings]
      summary: Get current owner's listings
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ListingDetail"
    post:
      operationId: createListing
      tags: [listings]
      summary: Create a new listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingDetail"
  /owner/listings/{id}:
    patch:
      operationId: updateListing
      tags: [listings]
      summary: Update a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListingDetail"
    delete:
      operationId: deleteListing
      tags: [listings]
      summary: Delete a listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── BOOKINGS ──────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    # ── Auth ──
...
    Listing:
      type: object
      required: [id, name, area, price, rating, review_count, img, amenities, status]
      properties:
        id:
          type: integer
        name:
          type: string
        area:
          type: string
        location:
          type: string
        price:
          type: integer
        type:
          type: string
        gender:
          type: string
        rating:
          type: number
        review_count:
          type: integer
        img:
          type: string
        amenities:
          type: array
          items:
            type: string
        badge:
          type: ["string", "null"]
        status:
          type: string
        desc:
          type: ["string", "null"]
    ListingDetail:
      type: object
      required: [id, name, area, price, rating, review_count, img, amenities, status]
      properties:
        id:
          type: integer
        name:
          type: string
        area:
          type: string
        location:
          type: string
        price:
          type: integer
        type:
          type: string
        gender:
          type: string
        rating:
          type: number
        review_count:
          type: integer
        img:
          type: string
        amenities:
          type: array
          items:
            type: string
        badge:
          type: ["string", "null"]
        status:
          type: string
        desc:
          type: ["string", "null"]
        deposit_months:
          type: integer
        meals_included:
          type: boolean
        wifi_included:
          type: boolean
        available_from:
          type: string
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        owner_id:
          type: integer
        owner_name:
          type: ["string", "null"]
        owner_since:
          type: ["string", "null"]
        images:
          type: array
          items:
            $ref: "#/components/schemas/ListingImage"
        created_at:
          type: string
    ListingsSearchResult:
      type: object
      required: [listings, total]
      properties:
        listings:
          type: array
          items:
            $ref: "#/components/schemas/Listing"
        total:
          type: integer
    ListingImage:
      type: object
      required: [id, url]
      properties:
        id:
          type: integer
        url:
          type: string
        caption:
          type: ["string", "null"]
        is_primary:
          type: boolean
    ListingInput:
      type: object
      required: [name, area, location, price, desc]
      properties:
        name:
          type: string
        area:
          type: string
        location:
          type: string
        price:
          type: integer
        type:
          type: string
        gender:
          type: string
        desc:
          type: string
        amenities:
          type: array
          items:
            type: string
        deposit_months:
          type: integer
        meals_included:
          type: boolean
        wifi_included:
          type: boolean
        available_from:
          type: string
        img:
          type: string
        badge:
          type: ["string", "null"]
    ListingUpdate:
      type: object
      properties:
        name:
          type: string
        area:
          type: string
        location:
          type: string
        price:
          type: integer
        type:
          type: string
        gender:
          type: string
        desc:
          type: string
        amenities:
          type: array
          items:
            type: string
        deposit_months:
          type: integer
        meals_included:
          type: boolean
        wifi_included:
          type: boolean
        available_from:
          type: string
        img:
          type: string
        badge:
          type: ["string", "null"]

    # ── Bookings ──
...